+++ libs/ngrx-toolkit/src/lib/redux-devtools/index.ts
@@ -1,0 +1,3 @@
+
+export * from './devtools-connect';
+export * from './with-devtools';
